Remove keyword constructor for strucrt StateMachine.
The reason for this change is somewhat complicated and relates to a known bug in method dispatch in Julia, which essentially causes constructor methods to become out of order in a testing environment due to different precompilation steps and shadowing each other. For now it remains easier to not use keyword constructors in the first place.