Refactor produce of Bus to use the storage potential return value of balance_on.
While the simple implementation of the energy system models for the various storage technologies do all have a capacity and load field, this might not be true for all cases and forever. It is better to use the interface balance_on provides to sum over the potential of connected systems on a bus.