Fix HeatPump and Electrolyser not working correctly as the first transformer in a chain.
This was caused because due to a performance-improving special case the first transformer in a chain of transformers will not the potential function called. The process function of the transformers however has several shrtcut checks. Running into these checks as the first transformer in a chain causes the component not to set any max_energy values, which are required for the following transformers to correctly calculate process().