Fix several components not considering energy potential in process and load steps.
This was previously not necessary as the balance component of energy exchanges had the correct values by itself, however with the new bus distribution concept it is now required to also check the energy potential. This also makes it necessary to prevent grid connections from receiving energy from other grid connections via the energy flow matrix.