Fix problem with a remaining error in energy calculations for PLRDEComponents.
The exact cause of the error is unknown, but appears to lie somewhere in the interpolation of energy values and/or efficiencies as the error appears quadratic to the distance to support values. The solution now is to also interpolate the energy values from the lookup tables instead of multiplying the power, PLR and efficiency at that PLR, which is more intuitive and faster for analytical efficiency functions.