Reimplement logic of HeatPump for calculation with other transformers.
This concerns the calculation when the other transformers have not had their potential step yet and the heat pump needs to considers all transformers as potential targets or sources for energy. This logic was removed erroneously as it is not necessary with other components that provide or take in infinite energy, such as grid connections.