Fix another problem with rejected passes of optimisation not resetting correctly.
The problem was again a change in the number of slices of a possible solution after a pass. The PLRs need to be reset back to the best guess after rejecting a slice. In this case a performance improvement can also be made in that we can quit optimisation when no further changes to the PLR are made because it would reject the same PLRs again and again until the maximum number of passes is reached.