Add test to catch a potential problem with inconsistent results of optimisation.
This problem does not occur with any currently implemented test, but we think it should occur in some cases, yet it doesn't. The test can be adapted to change the problem once it is clear what causes it and if it does. If the problem isn't a problem at all, the test simply tests that calculation happens as expected.