Fix BufferTank updating supply temperature between slices.
While it is a more realistic option for mixed or balanced models below the switch point, it causes problems because the component then writes in a higher max energy in its control step than it can actually supply in process. This fix solves that problem but introduces an inaccuracy as the entire energy supplied in a timestep is at the temperature before discharging.